pyproject.toml
[tool.poetry]
name = "fastapi-tutorial"
version = "1.0.0"
description = ""
authors = ["Gabriel Rockson <thegabrielrockson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
psycopg2-binary = "^2.9.3"
mypy = "^0.971"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
